rename testing.gui -> testing._gui and testing.eexception_handling -> testing._exception_handling
